backend/requirements.txt
fastapi
uvicorn
pydantic
tavily-python
python-dotenv
google-generativeai
markdown
reportlab
